flask_app/requirements.txt
flask
flask_sqlalchemy
flask-jwt-extended
werkzeug
pydantic